test page: making the new "skip to row N" link option work better.
before this commit, it didn't work very well w/ nvda.  b/c I wanted to "skip to row N" then press the down arrow and move nvda's cursor from there.  that didn't work.  the down arrow scrolled the viewport instead.  until I pressed capsLock+space.

my expectation was valid for a regular jump link on a regular page.  but not for the tricky JS-based jump-link-ish button that existed before this commit.

another thing I tried, then ruled out, was the code in it's current form plus a listener on the input text field for the ENTER key - so that pressing enter would do a click() on the link.  my expectation wasn't met in this case either.

so I gather nvda is doing something special when it follows a jump link, and I can't imitiate that with JS.